wrapper: add stub of a code generator for C APIs.
This is the start of an API code generator. Initially this
commit defines structs suitable for loading the libvirt API
definitions from their XML files, located using pkg-config.

It can be invoked using

  $ go generate

though at this commit it will merely parse the APIs and not
emit anything

It will default to loading the distro installed APIs, but
can be overridden using PKG_CONFIG_PATH to point to files
elsewhere. Most commonly though, it will be invoked via
the 'run' script in libvirt.git.

  $ ../libvirt.git/build/run go generate

Extracted from a larger commit

Signed-off-by: Victor Toso <victortoso@redhat.com>

Modified to

  * parse each API module separately, so that the output
    for QEMU/LXC layers can be separated from the core
    API.

  * eliminate CLI arg for overriding API file path, since
    setting $PKG_CONFIG_PATH can do this already, which
    is what the 'run' script does.

  * Wire up go:generate comment to run immediately,
    though currently it is a no-op beyond parsing the
    API.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>