wrapper: process enums to calculate raw values
The enum constants in the libvirt headers may be initialized
either with raw values, or with bitshifts, or with other enum
constants. In order to correctly sort them, we need to fully
resolve the raw values. A difficulty in doing this is that
some enums in the libvirt-qemu package are defined in terms
of enums in the libvirt package, so we need to resolve across
packages.

Extracted from a larger commit

Signed-off-by: Victor Toso <victortoso@redhat.com>

Modified to

  * Resolve the raw value as an integer, rather
    than resolve the hex value as a string

  * Handle resolving enums across packages, since
    each package is parsed separately.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>