Split off Wrapper function for each Helper function
The Wrapper functions are intended to be thin mappings over
the C API that adds a final virError parameter, to make it
easier to call from the Go layer.

Even though the Helper functions are in the C layer, we want
to have a consistent pattern for calling the libvirt APIs,
so we want the Helper functions to call Wrapper functions
just like the Go layer does. This will make it easier to
switch to auto-generating the Wrapper functions.

Extracted from a larger commit authored by:

  Victor Toso <victortoso@redhat.com>

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>