Fix some variable shadowing
In a few cases we're accidentally using ':=' instead of '=' in an
inner block, resulting in variable shadowing. As a consequence the
data we assign in the inner block doesn't propagate through the
rest of the function.

In the case of BlockRebase, FSThaw and FindStoragePoolSources the
result is that some parameters get left set to NULL when calling
the C API.

Fixes: https://gitlab.com/libvirt/libvirt-go-module/-/issues/12
Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>