connect: Allow obtaining raw connect pointer
I have a C library that has a function that accepts virConnectPtr
and I want to write Golang bindings for it. For instance:

  struct VirtLint *virt_lint_new(virConnectPtr conn);

Ideally, any consumer would use libvirt's Golang bindings to open
the connection and then just merely pass it to my library. But
since the raw connection pointer is not accessible outside of the
module, it's impossible to do so.

Therefore, introduce new RawPtr() function, that returns the raw
pointer. But since this is potentially dangerous (e.g. it can
mess up the global state we keep in @connections map) document
the function unsafe and discourage people from using it.

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>