wrapper: Alter name of dlopen function prototypes
The way our wrapper code generator works is the following:

1) Libvirt public API function pointer is declared, e.g.:

  typedef int
  (*virDomainAttachDeviceType)(virDomainPtr domain,
                               const char *xml);

  where the type name is constructed as "${funcName}Type".

2) A wrapper function is generated, e.g.

  int
  virDomainAttachDeviceWrapper(virDomainPtr domain.
                               const char *xml,
                               virErrorPtr err)

  which inside its body has a static variable:

  static virDomainGraphicsReloadType virDomainGraphicsReloadSymbol;

  which is then set to a corresponding libvirt API via dlopen().

This works beautifully, except it makes an assumption about
naming that no longer works in our favor. Specifically, another
part of our generator that handles generating typedefs for
function arguments (that are enums really, though uint
officially), uses the same naming scheme:

  ${funcName}${argName}

This is guaranteed to clash with each other when ${argName} =
"type" (the first letter of ${argName} is capitalized to follow
golang's naming scheme).

And we can already see an example of this in our code:

libvirt_generated_functions_dlopen_domain.go has:

  typedef int
  (*virDomainGraphicsReloadType)(virDomainPtr domain,
                                 unsigned int type,
                                 unsigned int flags);

and libvirt_generated_typedefs.h has:

  #if !LIBVIR_CHECK_VERSION(10, 2, 0)
  typedef int virDomainGraphicsReloadType;
  #endif

To make this kind of conflicts harder to hit, generate the
function type as:

  ${funcName}FuncType

and hope there won't be any API with 'funcType' as an argument
any time soon.

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>