Fix type of "MigrateDisksDetectZeroes" parameter
This parameter is a string list of disk names, not a boolean
flag.

Changing the type is an API break for any current user what is
setting this field, however, if they were setting this, then
libvirt would be failing with an error about the incorrect type
already.

To strictly preserve ABI though, we don't change the existing
field type, just document it as now unused, and add a second
parallel field with the correct type.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>